+++ src/TheNextWeek/box.h
@@ -19,8 +19,8 @@ class box: public hittable  {
     public:
         box() {}
         box(const vec3& p0, const vec3& p1, material *ptr);
-        virtual bool hit(const ray& r, float t0, float t1, hit_record& rec) const;
-        virtual bool bounding_box(float t0, float t1, aabb& box) const {
+        virtual bool hit(const ray& r, double t0, double t1, hit_record& rec) const;
+        virtual bool bounding_box(double t0, double t1, aabb& box) const {
                box =  aabb(pmin, pmax);
                return true; }
         vec3 pmin, pmax;
@@ -40,7 +40,7 @@ box::box(const vec3& p0, const vec3& p1, material *ptr) {
     list_ptr = new hittable_list(list,6);
 }
 
-bool box::hit(const ray& r, float t0, float t1, hit_record& rec) const {
+bool box::hit(const ray& r, double t0, double t1, hit_record& rec) const {
     return list_ptr->hit(r, t0, t1, rec);
 }
