+++ src/TheNextWeek/hittable_list.h
@@ -18,13 +18,13 @@ class hittable_list: public hittable  {
     public:
         hittable_list() {}
         hittable_list(hittable **l, int n) {list = l; list_size = n; }
-        virtual bool hit(const ray& r, float tmin, float tmax, hit_record& rec) const;
-        virtual bool bounding_box(float t0, float t1, aabb& box) const;
+        virtual bool hit(const ray& r, double tmin, double tmax, hit_record& rec) const;
+        virtual bool bounding_box(double t0, double t1, aabb& box) const;
         hittable **list;
         int list_size;
 };
 
-bool hittable_list::bounding_box(float t0, float t1, aabb& box) const {
+bool hittable_list::bounding_box(double t0, double t1, aabb& box) const {
     if (list_size < 1) return false;
     aabb temp_box;
     bool first_true = list[0]->bounding_box(t0, t1, temp_box);
@@ -42,7 +42,7 @@ bool hittable_list::bounding_box(float t0, float t1, aabb& box) const {
     return true;
 }
 
-bool hittable_list::hit(const ray& r, float t_min, float t_max, hit_record& rec) const {
+bool hittable_list::hit(const ray& r, double t_min, double t_max, hit_record& rec) const {
         hit_record temp_rec;
         bool hit_anything = false;
         double closest_so_far = t_max;
